chore: comments update, cleanup
