+++ src/fuzzer/analysis/typescript/ArgDef.ts
@@ -1,8 +1,4 @@
-import {
-  AST_NODE_TYPES,
-  parse,
-  simpleTraverse,
-} from "@typescript-eslint/typescript-estree";
+import { AST_NODE_TYPES } from "@typescript-eslint/typescript-estree";
 import {
   Identifier,
   TSTypeAnnotation,
@@ -12,12 +8,10 @@ import {
 import * as vscode from "vscode";
 
 /**
- * Provides basic Typescript analysis support for questions such as:
- *  1. What functions are in a module? --> findFnInSource()
- *  2. What arguments are in a function's signature?  --> getTsFnArgs()
- *
- * To answer question 2, an array of ArgDef objects is returned, which provides
- * information about the arguments.
+ * The ArgDef class describes a Typescript function argument using three input sources:
+ *  1. The argument's function signature --> type, dimension, optionality, offset
+ *  2. ArgOptions --> input intervals, how to handle any types
+ *  3. User overrides --> all values
  *
  * Argument types that are currently supported:
  * - Numbers
@@ -34,193 +28,6 @@ import * as vscode from "vscode";
  * - Non-primitive types
  * - Deconstructed types
  * - Generics
- *
- * Other limitations of the current implementation
- * - Requires a type-annotated TypeScript function signature
- * - Anonymous functions are not supported
- * - Analysis of class methods is not supported
- * - Presently cannot set an entire array as a constant (only its elements)
- * - String values are compared using default sort order, regarldess of the
- *   order specified in ArgOptions.strCharset.
- */
-
-/**
- * !!!
- */
-export class FunctionDef {
-  private argDefs: ArgDef<ArgType>[] = [];
-  private ref: FunctionRef;
-
-  // !!!
-  constructor(ref: FunctionRef, options?: ArgOptions) {
-    options = options ?? ArgDef.getDefaultOptions();
-    this.ref = ref;
-
-    this.argDefs = [];
-    const ast = parse(this.ref.src, { range: true }); // Parse the source
-
-    // Retrieve the function arguments
-    const fnDecl = ast.body[0];
-    if (
-      fnDecl.type === AST_NODE_TYPES.VariableDeclaration ||
-      fnDecl.type === AST_NODE_TYPES.FunctionDeclaration
-    ) {
-      const fnInit =
-        "declarations" in fnDecl ? fnDecl.declarations[0].init : fnDecl;
-      if (
-        fnInit !== null &&
-        (fnInit.type === AST_NODE_TYPES.FunctionExpression ||
-          fnInit.type === AST_NODE_TYPES.FunctionDeclaration ||
-          fnInit.type === AST_NODE_TYPES.ArrowFunctionExpression)
-      ) {
-        for (const i in fnInit.params) {
-          const thisArg = fnInit.params[i];
-          if (thisArg.type === AST_NODE_TYPES.Identifier) {
-            this.argDefs.push(
-              ArgDef.fromAstNode(thisArg, parseInt(i), options)
-            );
-          } else {
-            throw new Error(`Unsupported argument type: ${thisArg.type}`);
-          }
-        }
-      }
-    } else {
-      throw new Error(`Unsupported function declaration`);
-    }
-  }
-
-  // !!!
-  public getName(): string {
-    return this.ref.name;
-  }
-  // !!!
-  public getSrc(): string {
-    return this.ref.src;
-  }
-  // !!!
-  public getArgDefs(): ArgDef<ArgType>[] {
-    return this.argDefs;
-  }
-  // !!!
-  public getStartOffset(): number {
-    return this.ref.startOffset;
-  }
-  // !!!
-  public getEndOffset(): number {
-    return this.ref.endOffset;
-  }
-  // !!!
-  public getModule(): string {
-    return this.ref.module;
-  }
-  // !!!
-  public getRef(): FunctionRef {
-    return { ...this.ref };
-  }
-  // !!!
-  public applyOverrides(overrides: ArgOptionOverrides): void {
-    // Apply argument overrides
-    for (const argName of Object.keys(overrides.argOptions)) {
-      const arg = this.argDefs.find((arg) => arg.getName() === argName);
-      if (arg !== undefined) arg.setOptions(overrides.argOptions[argName]);
-    }
-  } // fn: applyOverrides()
-
-  /**
-   * Analyzes a Typescript module and returns a list of functions and their souces that
-   * are defined within the module.  If `fnName` and/or `offset` are specified, the list
-   * of functions will be filtered by the provided criteria.
-   *
-   * @param src The module source code to be analyzed
-   * @param fnName The optional function name to find
-   * @param offset The optional offset inside the desired function body
-   * @returns an array of `FunctionDef`s containing matching functions
-   *
-   * Throws an exception if the function is not supported for analysis.
-   */
-  public static find(
-    src: string,
-    module: string,
-    fnName?: string,
-    offset?: number,
-    options?: ArgOptions
-  ): FunctionDef[] {
-    let ret: FunctionRef[] = [];
-    const ast = parse(src, { range: true }); // Parse the source
-
-    // Traverse the AST to find function definitions
-    simpleTraverse(ast, {
-      enter: (node, parent) => {
-        if (
-          // Arrow Function Definition: const xyz = (): void => { ... }
-          node.type === AST_NODE_TYPES.VariableDeclarator &&
-          parent !== undefined &&
-          parent.type === AST_NODE_TYPES.VariableDeclaration &&
-          node.init &&
-          node.init.type === AST_NODE_TYPES.ArrowFunctionExpression &&
-          node.id.type === AST_NODE_TYPES.Identifier &&
-          (!fnName || node.id.name === fnName) &&
-          (!offset || (node.range[0] <= offset && node.range[1] >= offset))
-        ) {
-          ret.push({
-            name: node.id.name,
-            module: module,
-            src:
-              parent.kind + " " + src.substring(node.range[0], node.range[1]),
-            startOffset: node.range[0],
-            endOffset: node.range[1],
-          });
-        } else if (
-          // Standard Function Definition: function xyz(): void => { ... }
-          node.type === AST_NODE_TYPES.FunctionDeclaration &&
-          node.id !== null &&
-          (!fnName || node.id.name === fnName) &&
-          (!offset || (node.range[0] <= offset && node.range[1] >= offset))
-        ) {
-          ret.push({
-            name: node.id.name,
-            module: module,
-            src: src.substring(node.range[0], node.range[1]),
-            startOffset: node.range[0],
-            endOffset: node.range[1],
-          });
-        }
-        // TODO: Add support for class methods
-      }, // enter
-    }); // traverse AST
-
-    // If offset is provided and we have multiple matches,
-    // return the function that is closest to the offset
-    if (offset !== undefined && ret.length > 0) {
-      ret = [
-        ret.reduce((best, curr) =>
-          offset >= curr.startOffset &&
-          offset - curr.startOffset < offset - best.startOffset
-            ? curr
-            : best
-        ),
-      ];
-    }
-
-    return ret.map((e) => new FunctionDef(e, options));
-  } // fn: find
-
-  // !!!
-  public getArgDefsFlat(): ArgDef<ArgType>[] {
-    const ret: ArgDef<ArgType>[] = [];
-    for (const arg of this.argDefs) {
-      ret.push(arg);
-      ret.push(...arg.getChildrenFlat());
-    }
-    return ret;
-  } // fn: getArgDefsFlat()
-}
-
-/**
- * The ArgDef class describes a Typescript function argument using three input sources:
- *  1. The argument's function signature --> type, dimension, optionality, offset
- *  2. ArgOptions --> input intervals, how to handle any types
- *  3. User overrides --> all values
  */
 export class ArgDef<T extends ArgType> {
   private name: string; // name of the argument
@@ -675,7 +482,12 @@ export class ArgDef<T extends ArgType> {
     return [...this.children];
   }
 
-  // !!!
+  /**
+   * Returns a flat array of all arguments, including the children
+   * of arguments.  The selection is depth-first.
+   *
+   * @returns the argument's descendents (if it is an object)
+   */
   public getChildrenFlat(): ArgDef<ArgType>[] {
     const ret: ArgDef<ArgType>[] = [];
     for (const child of this.children) {
@@ -765,15 +577,6 @@ export class ArgDef<T extends ArgType> {
   } // fn: isOptionValid
 } // class: ArgDef
 
-// !!!
-export type FunctionRef = {
-  module: string;
-  name: string;
-  src: string;
-  startOffset: number;
-  endOffset: number;
-};
-
 /**
  * The set of options for an argument.  This option set is used to "fill in" information
  * that is not provided by analyzing the function.  For instance, a function signature
@@ -799,12 +602,16 @@ export type ArgOptions = {
   dftDimLength: Interval<number>; // Length of any dimension not specified in dimLength.
 };
 
-// !!!
+/**
+ * A set of option overrides for a set of arguments.
+ */
 export type ArgOptionOverrides = {
   [k: string]: ArgOptionOverride;
 };
 
-// !!!
+/**
+ * Argument option overrides
+ */
 export type ArgOptionOverride = {
   numInteger?: boolean;
   numSigned?: boolean;
