+++ src/fuzzer/generators/GeneratorFactory.test.ts
@@ -1,9 +1,5 @@
-import {
-  ArgDef,
-  ArgOptions,
-  ArgTag,
-  FunctionDef,
-} from "../analysis/Typescript";
+import { FunctionDef } from "../analysis/typescript/FunctionDef";
+import { ArgDef, ArgOptions, ArgTag } from "../analysis/typescript/ArgDef";
 import { setup, fuzz, FuzzOptions, getDefaultFuzzOptions } from "../Fuzzer";
 import { GeneratorFactory } from "./GeneratorFactory";
 import seedrandom from "seedrandom";
